Move FormValue to separate file
